% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docman.R
\name{bookdown_cleanup}
\alias{bookdown_cleanup}
\title{Clean up all files after a failed bookdown build}
\usage{
bookdown_cleanup(pspath = "cn", pspattern = "_main")
}
\arguments{
\item{pspath}{directory where the bookdown files are built}

\item{pspattern}{pattern matching the files to be deleted}
}
\description{
Whenever an error occurs during a build of a bookdown document
the cleanup procedure of RStudio does not delete all
output files. Files starting with _main
seam to survive the ordinary clean up process. This helper
function removes those files.
}
