% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{convertLibOToGraphic}
\alias{convertLibOToGraphic}
\title{Convert documents from source format into a given target output format}
\usage{
convertLibOToGraphic(psLibOFile, psOutFormat = NULL, psLibODir = "odg",
  psFigOutDir = ".")
}
\arguments{
\item{psLibOFile}{name of the libre office graphics file}

\item{psLibODir}{source directory of Libre Office files}

\item{psFigOutDir}{output directory where figure pdfs are expected to be}
}
\description{
\code{convertLibOToGraphic} assumes that LibreOffice is installed
and is ideally available on the search path. Because more recent
versions of windows and mac osx appear to have problems with that
we started to use absolute paths to LibO installations. Source files
which are assumed to LibO-draw files in .odg-format are converted on the
fly to the specified output format which are then included in the
source R markdown document
}
