% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{create_beamer_presentation}
\alias{create_beamer_presentation}
\title{Create a beamer presentation Rmarkdown (Rmd) document}
\usage{
create_beamer_presentation(psDocuName, psPkgPath = ".",
  psRmdTemplate = "default_pdf_beamer_presentation",
  psTemplatePkg = "rmddochelper", psDocuSubdir = "vignettes",
  pDocuHasOwnSubdir = TRUE, pbOverwrite = FALSE, pbEdit = TRUE,
  plReplace = NULL)
}
\arguments{
\item{psPkgPath}{path where package is located under which document should be created}

\item{psRmdTemplate}{name of the template to be used}

\item{psTemplatePkg}{package from where the template should be taken}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbOverwrite}{flag whether existing files are overwritten}

\item{pbEdit}{directly open newly created document}

\item{plReplace}{list with replacement key-values}

\item{pbDocuHasOwnSubdir}{should document be stored in separate subdir}
}
\description{
The function \code{create_beamer_presentation} is a wrapper
to create_docu_skeleton using template default_pdf_beamer_presentation
to generate a skeleton Rmarkdown (Rmd) document. The skeleton
can be used to generate a beamer presentation.

#' @param   psDocuName           name of the new document
}
