% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{create_docu_skeleton}
\alias{create_docu_skeleton}
\title{Create a Rmarkdown (Rmd) document based on the provided template}
\usage{
create_docu_skeleton(psDocuName, psPkgPath = ".", psRmdTemplate,
  psTemplatePkg = NULL, psDocuSubdir = "vignettes",
  pDocuHasOwnSubdir = TRUE, pbOverwrite = FALSE, pbEdit = TRUE)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPkgPath}{path where package is located under which document should be created}

\item{psRmdTemplate}{name of the template to be used}

\item{psTemplatePkg}{package from where the template should be taken}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbOverwrite}{flag whether existing files are overwritten}

\item{pbEdit}{directly open newly created document}

\item{pbDocuHasOwnSubdir}{should document be stored in separate subdir}
}
\description{
The \code{rmarkdown} package provides a templating mechanism
within which templates can be stored in the source version
of a package in directory inst/rmarkdown/templates. By default,
we choose the template from the rmarkdown package for rmd
documents

Create a skeleton for a new Rmarkdown (Rmd) document

\code{create_docu_skeleton} assumes that psPkgPath is a
directory that contains an R-package. By default the new
document is created in subdirectory "vignettes". If this
subdirectory does not exist, it is created. The document
is generated by the function \code{rmarkdown::draft} using
the template "project_docu".
}
\details{
The basic functionality follows the function
\code{devtools::use_vignette}, except for the possibility
of specifying any given template from any package.
}

