% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{create_html_output_skeleton}
\alias{create_html_output_skeleton}
\title{Create a Rmarkdown (rmd) document with html_output}
\usage{
create_html_output_skeleton(psDocuName, psPkgPath = ".",
  psRmdTemplate = "default_html_output", psTemplatePkg = "rmddochelper",
  psDocuSubdir = "vignettes", pDocuHasOwnSubdir = TRUE,
  pbOverwrite = FALSE, pbEdit = TRUE, plReplace = NULL)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPkgPath}{path where package is located under which document should be created}

\item{psRmdTemplate}{name of the template to be used}

\item{psTemplatePkg}{package from where the template should be taken}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbOverwrite}{flag whether existing files are overwritten}

\item{pbEdit}{directly open newly created document}

\item{plReplace}{list with replacement key-values}

\item{pbDocuHasOwnSubdir}{should document be stored in separate subdir}
}
\description{
\code{create_html_output_skeleton} is a wrapper to
\code{create_docu_skeleton} for easily creating html-documents
with reasonable default settings.
}

