% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{create_odg_graphic}
\alias{create_odg_graphic}
\title{Create an empty odg graphic}
\usage{
create_odg_graphic(psGraphicName = "skeleton.odg",
  psOdgTemplate = "odg_figure", psTemplatePkg = "rmddochelper",
  pbEdit = TRUE)
}
\arguments{
\item{psGraphicName}{Format of diagram to be created}
}
\description{
\code{create_odg_graphic} uses the empty template skeleton.odg which
is stored in the rmarkdown template directory of this package.
When calling  \code{create_odg_graphic} without any arguments, then the
template is opened using soffice draw. When a filename for the odg graphic
which should be created, is specified the template is renamed to the given
name.
}

