% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{create_pdf_output_skeleton}
\alias{create_pdf_output_skeleton}
\title{Create a Rmarkdown (Rmd) document with pdf_output}
\usage{
create_pdf_output_skeleton(psDocuName, psPkgPath = ".",
  psRmdTemplate = "default_pdf_output", psTemplatePkg = "rmddochelper",
  psDocuSubdir = "vignettes", pDocuHasOwnSubdir = TRUE,
  pbOverwrite = FALSE, pbEdit = TRUE)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPkgPath}{path where package is located under which document should be created}

\item{psRmdTemplate}{name of the template to be used}

\item{psTemplatePkg}{package from where the template should be taken}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbOverwrite}{flag whether existing files are overwritten}

\item{pbEdit}{directly open newly created document}

\item{pbDocuHasOwnSubdir}{should document be stored in separate subdir}
}
\description{
Function {create_pdf_output_skeleton} is a wrapper to
\code{create_docu_skeleton} for easily creating pdf-documents
with reasonable default settings.
}

