% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_project_management.R
\name{delete_files}
\alias{delete_files}
\title{Helper function to delete a set of files}
\usage{
delete_files(ps_docu_path, pvec_filenames, pb_interactive = TRUE)
}
\arguments{
\item{pvec_filenames}{vector of filenames to be deleted}

\item{pb_interactive}{flag for interactive mode}
}
\description{
If \code{pb_interactive} is true, the function loops
over a given vector of file names and asks to user
whether the files should be deleted. Otherwise the
files in the vector are deleted without asking.
}
