% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_project_management.R
\name{find_match_delete}
\alias{find_match_delete}
\title{Helper to find all files match a certain pattern}
\usage{
find_match_delete(ps_docu_path, ps_pattern, pb_ignore_dir = TRUE,
  pb_interactive = TRUE)
}
\arguments{
\item{ps_docu_path}{directory containing files to be matched}

\item{ps_pattern}{pattern used for matching files}

\item{pb_ignore_dir}{flag indicating whether directories are ignored}

\item{pb_interactive}{flag for interactive mode}
}
\description{
A given pattern is used in function \code{list.files()}
to obtain a list of files matching the given pattern.
The list of files is then used as input for the function
\code{delete.files()}.
}
