% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{genericScaledPlot}
\alias{genericScaledPlot}
\title{Insert a plot into a document using a given width scale}
\usage{
genericScaledPlot(pData = NULL, pnPaperWidthScale, pfPlotMethod = NULL,
  ...)
}
\arguments{
\item{pData}{data to be plotted}

\item{pnPaperWidthScale}{scale factor for graphics width}

\item{pfPlotMethod}{function that should produce the plot}
}
\description{
\code{genericScaledPlot} is a work-around due to the issue with setkeys{Gin}
which, we have to use because of the background picture in the header and which
has to be specified before each plot because otherwise, the plot is not visible.
}
\details{
The function \code{genericScaledPlot} can also be used to just produce the
LaTeX statement that sets the Gin key. This can be done when leaving
pfPlotMethod NULL.
}
