% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{get_repl_tag}
\alias{get_repl_tag}
\title{Convert a tag to a replacement placeholder}
\usage{
get_repl_tag(psTag)
}
\arguments{
\item{psTag}{tag to be converted to a placeholder}
}
\value{
replacement placeholder
}
\description{
Convert a tag to a replacement placeholder
}

