% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docman.R
\name{get_this_rmd}
\alias{get_this_rmd}
\title{Get the name and the path to the current rmd-file}
\usage{
get_this_rmd(ps_wd = getwd())
}
\arguments{
\item{ps_wd}{full path of current working directory}
}
\value{
name and full path to current rmd-source file.
}
\description{
Analogously, to \code{rprojroot::this_file()}, we want to determine
the name and the path to the currently active rmd-source document.
The solution with \code{rprojroot::this_file()} works only, when the
document is knitted, but not when run from inside an R-chunk.
Given the name and the path a a rmd-source document project
follow a certain pattern, the full path and the name of an
rmd-source document can be determined based on the result
of the function \code{getwd()}.
}
