% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{gsub_templ_pattern}
\alias{gsub_templ_pattern}
\title{Replacement of placeholder by replacement list values}
\usage{
gsub_templ_pattern(ps_docu_template, pl_repl_info)
}
\arguments{
\item{ps_docu_template}{document template string}

\item{pl_repl_info}{replacement info list}
}
\value{
s_docu_result      document result string
}
\description{
Replacement of placeholder by replacement list values
}

