% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{has_file_search_pat}
\alias{has_file_search_pat}
\title{Check whether ps_pattern is found in ps_file}
\usage{
has_file_search_pat(ps_file, ps_pattern)
}
\arguments{
\item{ps_file}{name of file in which search is done}

\item{ps_pattern}{pattern for which we search for}
}
\value{
TRUE, if we have an exact match of ps_pattern in ps_file, FALSE otherwise
}
\description{
The check is done via grep for the pattern in the
file. First, the file content is read into a character
vector, then we search for the pattern. We are
only interested in exact matches, hence the argument
fixed=TRUE is specified for grep().
}
