% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{include_graphics_cmd}
\alias{include_graphics_cmd}
\title{Add statement to include graphic into Rmarkdown source}
\usage{
include_graphics_cmd(psGraphicName, pvRmdSrc, psGrFmt)
}
\arguments{
\item{psGraphicName}{name of the graphic file to be included}

\item{pvRmdSrc}{name of the Rmarkdown source file}

\item{psGrFmt}{graphic format}
}
\value{
vRmdSrc        vector with extended Rmarkdown sources
}
\description{
\code{include_graphics_cmd} assumes that in the Rmarkdown source
file a chunk is inserted which will contain the statement that
includes the graphic file. The chunk needs to be labelled with
the name of the graphic file to be included. This function searches
for this chunk and inserts the statement. If the chunk is not
found, the graphics statement is included above the end of
document marker.
}
