% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{includeOdgGraphic}
\alias{includeOdgGraphic}
\title{Include a graphic file based on an odg template in a given output format

\code{includeOdgGraphic} takes the name of a file containing a graphic
in odg format, converts the content of that file into a given output format using
function \code{convertLibOToGraphic} and outputs the string in markdown
format to include the figure. Graphic files are only re-generated
if the outputfile does not exist or, if the flag pbMustGenerate is TRUE.}
\usage{
includeOdgGraphic(psOdgFileStem, psOutFormat, psOdgDir = "odg",
  psFigOutDir = ".", pbMustGenerate = FALSE, pnPaperWidthScale = NULL)
}
\arguments{
\item{psOdgFileStem}{stem of odg figure file}

\item{psOutFormat}{output format of the graphic file to be included}

\item{psOdgDir}{directory where odg figure file is stored}

\item{psFigOutDir}{directory where output should be placed}

\item{pbMustGenerate}{flag to indicate whether pdf-graphics must be regenerated}

\item{pnPaperWidthScale}{scale factor for produced graphic}
}
