% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{includeOdsTable}
\alias{includeOdsTable}
\title{Generic Method to include Ods tables as graphics}
\usage{
includeOdsTable(psOdsFileStem, psOutFormat = "pdf", psOdsDir,
  psFigOutDir = ".", pbMustGenerate = FALSE,
  pnPaperWidthScale = NULL)
}
\arguments{
\item{psOdsFileStem}{stem of odg figure file}

\item{psOutFormat}{output format of the graphic file to be included}

\item{psOdsDir}{directory where odg figure file is stored}

\item{psFigOutDir}{directory where output should be placed}

\item{pbMustGenerate}{flag to indicate whether pdf-graphics must be regenerated}

\item{pnPaperWidthScale}{scale factor for produced graphic}
}
\description{
This function works the same as includeOdgGraphic, but
it converts LibreOffice Tables from ods-format to pdf.
}
