% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{insertOdgAsPdf}
\alias{insertOdgAsPdf}
\title{Inserts an odg graphic in pdf format into a rmarkdown text}
\usage{
insertOdgAsPdf(psOdgFileStem, psOdgDir = "odg", psFigOutDir = ".",
  pbMustGenerate = FALSE, pnPaperWidthScale = NULL)
}
\arguments{
\item{psOdgFileStem}{stem of odg figure file}

\item{psOdgDir}{directory where odg figure file is stored}

\item{psFigOutDir}{directory where output should be placed}

\item{pbMustGenerate}{flag to indicate whether pdf-graphics must be regenerated}
}
\description{
This function is a wrapper to the more generic function
\code{includeOdgGraphic} which allows to include graphics
into rmarkdown source files.
}

