% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{insertOdgAsPng}
\alias{insertOdgAsPng}
\title{Include an odg graphic in png format into a rmarkdown source file}
\usage{
insertOdgAsPng(psOdgFileStem, psOdgDir = "odg", psFigOutDir = ".",
  pbMustGenerate = FALSE, pnPaperWidthScale = NULL)
}
\arguments{
\item{psOdgFileStem}{stem of odg figure file}

\item{psOdgDir}{directory where odg figure file is stored}

\item{psFigOutDir}{directory where output should be placed}

\item{pbMustGenerate}{flag to indicate whether pdf-graphics must be regenerated}

\item{pnPaperWidthScale}{scale factor for produced graphic}
}
