% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{insertOdsAsPdf}
\alias{insertOdsAsPdf}
\title{Insert an LibreOffice table into a Rmd document}
\usage{
insertOdsAsPdf(psOdsFileStem, psOdsDir = "ods", psFigOutDir = ".",
  pbMustGenerate = FALSE, pnPaperWidthScale = NULL)
}
\arguments{
\item{psOdsFileStem}{stem of ods file}

\item{psOdsDir}{directory where ods file is stored}

\item{psFigOutDir}{directory where output should be placed}

\item{pbMustGenerate}{flag to indicate whether pdf-graphics must be regenerated}

\item{pnPaperWidthScale}{scale factor for produced graphic}
}
\description{
Insert an LibreOffice table into a Rmd document
}

