% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helper.R
\name{matGetMatElem}
\alias{matGetMatElem}
\title{Matrix with character elements and column and row indices}
\usage{
matGetMatElem(psBaseElement, pnNrRow, pnNrCol)
}
\arguments{
\item{psBaseElement}{character prefix shown in front of indices}

\item{pnNrRow}{number of rows}

\item{pnNrCol}{number of columns}
}
\description{
Matrix with character elements and column and row indices
}
