% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_project_management.R
\name{move_doc_to_subdir}
\alias{move_doc_to_subdir}
\title{Move a document to a separate subdirectory}
\usage{
move_doc_to_subdir(psDocuName, psDocuPath = NULL, pbMoveEverything = TRUE)
}
\arguments{
\item{psDocuName}{Name of the document to be moved}

\item{psDocuPath}{Path where psDocuName is stored}

\item{pbMoveEverything}{flag whether complete content of current directory should be moved to new directory}
}
\description{
For a given document specified by parameter psDocuName
create a new subdirectory and move the document and
everything belonging to that document to the new
subdirectory. If the parameter pbMoveEverything is
set to TRUE, then not only the document itself,
but the whole content of the current directory is
moved. The reason for that is that TeX-based documents
can be spread over multiple directories and can
include information from other files.
}
\details{
The path to the document file to be moved can be specified
either via the parameter psDocuPath or it can also be
included in the parameter psDocuName. But if the latter
option is chosen, then psDocuPath must be NULL.
}

