% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{odg.graphics.conv.hook}
\alias{odg.graphics.conv.hook}
\title{Hook function for odg-graphics conversion}
\usage{
odg.graphics.conv.hook(before, options, envir)
}
\arguments{
\item{before}{flat to indicate which statements must be executed before the chunk}

\item{options}{list of options passed from the chunk header to the hook function}

\item{envir}{environment}
}
\description{
\code{odg.graphics.conv.hook} is a hook function for knitr chunks which include
given graphics using function \code{knitr::include_graphics()}. Whenever graphics
are produced using LibreOffice Draw and saved as odg-files this hook function can
be used to automatically convert the odg-graphics files into pdf-files.
}

