% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{odg_convert_hook}
\alias{odg_convert_hook}
\title{Conversion Hook for Odg Graphics Objects}
\usage{
odg_convert_hook(before, options, envir)
}
\arguments{
\item{before}{running before chunk code}

\item{options}{chunk label options}

\item{envir}{environment}
}
\value{
command to include the converted graphics object files
}
\description{
A graphics object in an odg-formatted file cannot be
included directly into an Rmarkdown source document.
The odg-file must first be converted into a format
such as pdf or png. This conversion is done automatically
by this hook-function whenever a R-code-chunk with a
statement to include a graphics object is executed.
}
