% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{odg_draft}
\alias{odg_draft}
\title{Copy a draft template file for a odg graphics}
\usage{
odg_draft(file, template = "odg_figure", package = NULL,
  create_dir = "default", pbOverwrite = FALSE)
}
\arguments{
\item{file}{name of the new document}

\item{template}{name of the template}

\item{package}{package where template can be found}

\item{create_dir}{whether or not to create a new directory for this document}

\item{pbOverwrite}{should existing files be overwritten}

\item{plReplace}{list with replacement key-values}
}
\value{
file          name of the new document
}
\description{
This function \code{odg_draft} works analogously to
\code{rmarkdown::draft}, but for ODG graphics files
}

