% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_r6class.R
\docType{class}
\name{R6ClassEnum}
\alias{R6ClassEnum}
\title{R6 Class representing enumerations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassEnum$new()
}
\description{
Simple R6 class allowing to include enumeration objects
in documents. The document formats must allow for R-code-junks
to be executed, as e.g. with rmarkdown or rnoweb documents.
The enumeration objects are initialized at the beginning
of the document and whenever an enumeration object is needed
an inline R-statement is used to generate the required enumeration
instance.
}
\section{Fields}{

\describe{
\item{\code{count}}{enumeration count}
}}

\section{Public methods}{

\describe{
  \item{\code{new()}}{Instantiation of enumeration object}
  \item{\code{setCount(pnCount)}}{setter for count field}
  \item{\code{getCount()}}{getter for count field}
  \item{\code{incrCount()}}{increment count by one}
  \item{\code{incrCountByN(pnIncr)}}{increment count by pnIncr}
  \item{\code{out()}}{output count}
}
}

\examples{
# generate a new enumeration object
en <- rmddochelper::R6ClassEnum$new()
# increment number
en$incrCount()
# output of count
en$out()
}
\keyword{datasets}
