% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_table_abbrev_r6class.R
\docType{data}
\name{R6ClassTableAbbrev}
\alias{R6ClassTableAbbrev}
\title{R6 Class To Model Abbreviation Table Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassTableAbbrev
}
\description{
Abbreviation tables are tables with two columns one of which
contains the abbreviations and the second column contains the
meaning of the abbreviations. The tables are specific for a
given document. The R6 class defined here provides functionlity
to automatically generate a table of abbreviation
}
\keyword{datasets}

