% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_doc_r6class.R
\docType{class}
\name{R6RmdDocInfo}
\alias{R6RmdDocInfo}
\title{R6 Class Representing Rmarkdown Document Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6RmdDocInfo$new()
}
\value{
R6 Object of type \code{\link{R6RmdDocInfo}}
}
\description{
Rmarkdown documents which include information that is
distributed over several files are ideally put into
a single parent directory. For organisational reasons,
it is easier to put the different files included in
a Rmarkdown document into different subfolders according
to their file-type. This structure is constant, except
for the name of the Rmarkdown file which can easily be
parameterized.
}
\details{

}
\section{Fields}{

\describe{
\item{\code{docName}}{Name of the document project}

\item{\code{rmdPath}}{Path where rmd-files are stored}

\item{\code{figPath}}{Path where figure source files are stored}

\item{\code{pngPath}}{Path where figure files that should be included in the document are stored}
}}

\section{Public methods}{

\describe{
  \item{\code{new()}}{Instantiation of R6 object of class R6RmdDocInfo}
  \item{\code{setDocName(psDocName)}}{setter for document project name}
  \item{\code{getDocPath}}{getter for field docName}
  \item{\code{setRmdPath(psRmdPath)}}{setter for field rmdPath}
  \item{\code{getRmdPath}}{getter for field rmdPath}
  \item{\code{setFigPath(psFigPath)}}{setter for field figPath}
  \item{\code{getFigPath}}{getter for field figPath}
  \item{\code{setPngPath(psPngPath)}}{setter for field pngPath}
  \item{\code{getPngPath}}{getter for field pngPath}
}
}

\examples{
r6objDocInfo <- R6RmdDocInfo$new(psDocName = "MyFirstRmdDoc")
}
\keyword{datasets}
