% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_skeleton_creator.R
\name{replace_placeholders}
\alias{replace_placeholders}
\title{Replacement of placeholders in skeletonfile}
\usage{
replace_placeholders(psTrgFile, plReplace)
}
\arguments{
\item{psTrgFile}{path to and including name of targetfile}

\item{plReplace}{replacement list}
}
\description{
Placeholders in skeleton files are replaced using gsub
functionalities
}
