% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper.R
\docType{package}
\name{rmddochelper}
\alias{rmddochelper}
\alias{rmddochelper-package}
\title{rmddochelper: A package containing helper functions for rmarkdown documents.}
\description{
The rmddochelper package provides the following functions
\describe{
  \item{\code{create_pdf_output_skeleton}}{create a skeleton document for pdf output}
  \item{\code{create_odg_graphic}}{create a skeleton odg graphic}
  \item{\code{insertOdgAsPdf}}{convert an odg graphic to pdf format and insert it in a pdf document}
  \item{\code{insertOdgAsPng}}{convert an odg graphic to png format and insert it in a html document}
  \item{\code{cleanup_output}}{cleanup of temporary document compilation output}
}
}
\section{R6 classes}{

Document status and table of abbreviations are meta-type of information about the content of the document.
Updating this type of information can be difficult, because it needs additional attention besides the content
of the document. To reduce the extra burden of maintaining the extra information on document status and
abbreviations, the two R6 classes are designed to automate the maintenance of this information as much as
possible.
}

\section{Templates}{

Besides the functions and the R6 classes, we have also created a number of templates that come with this
package. These are very generic templates that are used when creating a document.
}

