% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_helper.R
\name{sGetParentRow}
\alias{sGetParentRow}
\title{Determine parent row in demo of Meuwissen-Luo algorithm}
\usage{
sGetParentRow(pPed, pbShowBothNaParents = TRUE)
}
\arguments{
\item{pPed}{the pedigree}

\item{pbShowBothNaParents}{flag whether to show NA parents when both parents are unknown}
}
\value{
a character vector with the same length as the pedigree containing the parents
}
\description{
Given a pedigree the demo of determining the relationship
matrix using Meuwissen-Luo writes the parents above the
label-row of the relationship matrix
}
