% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{use_odg_graphic}
\alias{use_odg_graphic}
\title{Add ODG Graphic To A Document}
\usage{
use_odg_graphic(ps_odg_file, ps_odg_template = "odg_figure",
  ps_template_package = "rmddochelper", pb_recursive = TRUE,
  pb_edit = TRUE)
}
\arguments{
\item{ps_odg_file}{file name including path of ODG graphic}
}
\description{
Similar to the \code{devtools::use_*} functions, we add an ODG
graphic file to a document. The file to be added is copied from
a template which is stored in this package. After copying the
template, we have to modify the template which is done by a
pre-defined program or tool. Once we are done with the modification,
the ODG graphic is converted into a pre-defined set of output
formats. Those output files will then be included in the document.
}
\examples{
\dontrun{
rmddochelper::use_odg_graphic(ps_odg_file = "my_odg_graphic.odg")
}

}
