% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtools.R
\name{write_bib}
\alias{write_bib}
\title{Write bibentries to a bib file}
\usage{
write_bib(pvec_bref, ps_bibfile = "skeleton.bib")
}
\arguments{
\item{pvec_bref}{vector of bibentries}

\item{ps_bibfile}{bibfile where bibentries should be written to}
}
\description{
One way of specifying records of a bibliography is to define
a vector of bibentries at the end of a document. This vector
must then be written to a .bib file which can later be used
by the latex-system to generate the bibliography. This function
is used to write the vector of bibentries to a file.
}
