% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_project_management.R
\name{cleanup_all}
\alias{cleanup_all}
\title{Recursively cleanup all directories of psDocuPath}
\usage{
cleanup_all(psDocuPath = "vignettes", psPattern = c("pdf$", "png$"),
  pbIgnoreDir = TRUE)
}
\arguments{
\item{psDocuPath}{Name of the document path}

\item{psPattern}{Pattern to match by files to be deleted}
}
\description{
This is a wrapper over cleanup_output using sapply
over the result of list.files(psDocuPath)
}
