% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_project_management.R
\name{cleanup_output}
\alias{cleanup_output}
\title{Cleanup a directory from temporary output files}
\usage{
cleanup_output(psDocuPath = "vignettes", psPattern = c("pdf$", "png$"),
  pbIgnoreDir = TRUE, pbInteractive = TRUE)
}
\arguments{
\item{psDocuPath}{Name of the document path}

\item{psPattern}{Pattern to match by files to be deleted}

\item{pbIgnoreDir}{Flag indicating whether directories should be ignored}

\item{pbInteractive}{Should user be asked whether to delete files}
}
\description{
Remove all files and possibly also directories that
were generated as ouput from a previous compilation
of a document in directory psDocuPath. All files that
match the pattern specified by psPattern are selected
for deletion. The user is asked whether the files should
be deleted before actually deleting them. By default,
directories are ignored, even if they match the given
pattern. In case directories are to be included the
argument pbIgnoreDir must be set to FALSE.
}

