% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_project_management.R
\name{cleanup_output}
\alias{cleanup_output}
\title{Cleanup a directory from old output}
\usage{
cleanup_output(psDocuName = NULL, psDocuPath = "vignettes",
  psPattern = c("pdf$", "png$"), pbIgnoreDir = TRUE)
}
\arguments{
\item{psDocuName}{Name stem of the directory source}

\item{psDocuPath}{Name of the document path}

\item{psPattern}{Pattern to match by files to be deleted}
}
\description{
Remove all files and directories that were generated
as ouput from a previous compilation of a document
given by psDocuName and match the pattern specified by
psPattern. By default, directories are ignored,
even if they match the given pattern.
}
\details{
In case psDocuName is left NULL then all subdirectories of
psDocuPath are searched for files to be removed
}

