% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{create_odg_graphic}
\alias{create_odg_graphic}
\title{Create an empty odg graphic}
\usage{
create_odg_graphic(psGraphicName = "skeleton.odg",
  psGraphicPath = "vignettes", psRmdSrcFile, psGrFmt = "pdf",
  psOdgTemplate = "odg_figure", psTemplatePkg = "rmddochelper",
  create_dir = "default", pbRecursive = TRUE, pbEdit = TRUE)
}
\arguments{
\item{psGraphicName}{Format of diagram to be created}

\item{psGraphicPath}{Path where created odg file should be stored}

\item{psOdgTemplate}{name of the template to be used}

\item{psTemplatePkg}{package where template is stored}

\item{create_dir}{should created odg file be stored in separate directory}

\item{pbRecursive}{recursively create complete path to graphic file}

\item{pbEdit}{directly edit created odg file}
}
\description{
\code{create_odg_graphic} uses templates which are either provided
by the caller or which are stored in the rmarkdown/template directory
of this package `rmddochelper`. When calling  \code{create_odg_graphic}
without any arguments, then the template is opened using soffice draw.
When a filename for the odg graphic which should be created, is
specified the template is renamed to the given name.
}

