% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{create_odg_graphic}
\alias{create_odg_graphic}
\title{Create an empty odg graphic}
\usage{
create_odg_graphic(psGraphicName = "skeleton.odg",
  psGraphicPath = "vignettes", psRmdSrcFile = NULL, psGrFmt = "pdf",
  psOdgTemplate = "odg_figure", psTemplatePkg = "rmddochelper",
  create_dir = "default", pbRecursive = TRUE, pbEdit = TRUE,
  psChunkLabel = "echo=FALSE, odg.conv=TRUE, odg.path=\\"odg\\", odg.graph.cache=TRUE")
}
\arguments{
\item{psGraphicName}{Format of diagram to be created}

\item{psGraphicPath}{Path where created odg file should be stored}

\item{psRmdSrcFile}{rmd source file where include statement is inserted}

\item{psGrFmt}{graphics format}

\item{psOdgTemplate}{name of the template to be used}

\item{psTemplatePkg}{package where template is stored}

\item{create_dir}{should created odg file be stored in separate directory}

\item{pbRecursive}{recursively create complete path to graphic file}

\item{pbEdit}{directly edit created odg file}

\item{psChunkLabel}{string with additional chunk labels}
}
\description{
\code{create_odg_graphic} uses templates which are either provided
by the caller or which are stored in the rmarkdown/template directory
of this package `rmddochelper`. When calling  \code{create_odg_graphic}
without any arguments, then the template is opened using soffice draw.
When a filename for the odg graphic which should be created, is
specified the template is renamed to the given name.
}
\details{
If the parameter psRmdSrcFile is specified, a command to include the
created graphic file in the specified format is inserted into the
rmd-source file. If psRmdSrcFile is left at its default (null) then
no include command is inserted. The command to include graphics
is based on the function \code{knitr::include_graphics()}. When
specifying the parameter psChunkLabel an additional chunk label
can be specified.
}
