% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{insertOdgAsPdf}
\alias{insertOdgAsPdf}
\title{Inserts an odg draw graphic into a rmarkdown text}
\usage{
insertOdgAsPdf(psOdgFileStem, psOdgDir = "odg", psFigOutDir = ".",
  pbMustGenerate = FALSE, pnPaperWidthScale = NULL)
}
\arguments{
\item{psOdgFileStem}{stem of odg figure file}

\item{psOdgDir}{directory where odg figure file is stored}

\item{psFigOutDir}{directory where output should be placed}

\item{pbMustGenerate}{flag to indicate whether pdf-graphics must be regenerated}
}
\description{
\code{insertOdgAsPdf} takes the name of a file containing a graphic
in odg format, converts the content of that file into pdf using
function \code{convertLibOToPdf} and outputs the string in markdown
format to include the figure. Pdf-formatted graphics are only re-generated
if the pdf-file does not exist or, if the flag pbMustGenerate is TRUE.
}

