% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{knitr_include_graphics_pdf}
\alias{knitr_include_graphics_pdf}
\title{Add statement to include graphic via knitr::include_graphics()}
\usage{
knitr_include_graphics_pdf(psGraphicName, pvRmdSrc)
}
\arguments{
\item{psGraphicName}{name of the graphic file to be included}

\item{pvRmdSrc}{name of the Rmarkdown source file}
}
\value{
vector with extended Rmarkdown sources
}
\description{

}

