% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_odg_graphics.R
\name{odg.graphics.conv.hook}
\alias{odg.graphics.conv.hook}
\title{Hook function for odg-graphics conversion}
\usage{
odg.graphics.conv.hook(before, options, envir)
}
\arguments{
\item{before}{flat to indicate which statements must be executed before the chunk}

\item{options}{list of options passed from the chunk header to the hook function}

\item{envir}{environment}
}
\description{
\code{odg.graphics.conv.hook} is a hook function for knitr chunks which include
given graphics using function \code{knitr::include_graphics()}. Whenever graphics
are produced using LibreOffice Draw and saved as odg-files this hook function can
be used to automatically convert the odg-graphics files into files of a different format.
}
\details{
The output format is either set by default to pdf or it is taken from options$out.format.
In the current version no validity checks for the output format is done. Whatever can be
handled by LibreOffice which is doing the conversion is fine.

Please have a look at the examples section, to see how this hook function can be used. At
least two things must be done. First the hook function must be registered using the
function \code{knitr::knit_hooks$set()} and secondly the code junk must activate the
usage of the hook-function by setting the label used in the activation function to TRUE.

Paths in the chunk-labels are either absolute or relative to the location of the Rmd-file
}
\examples{
\dontrun{
# activation of hook function somewhere at the top of the document
```{r setup, include=FALSE}
knitr::knit_hooks$set(odg.conv = rmddochelper::odg.graphics.conv.hook)
```
# ... later in the document use the hook function in a junk
```{r some_graphic, odg.conv = TRUE, odg.path="<path-to-odg-files", odg.out.dir="path-to-odg.out.dir"}
knitr::include_graphics(path = "<path-to-odg.out.dir>/some_graphic")
```
 }
}
