% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_table_r6class.R
\docType{class}
\name{R6ClassDocuStatus}
\alias{R6ClassDocuStatus}
\title{R6 Class Representing Document Status Objects}
\format{\code{\link{R6Class}} object.}
\usage{
R6ClassDocuStatus$new()
}
\value{
Object of \code{\link{R6Class}} with methods for managing the status of a document.
}
\description{
\code{R6ClassDocuStatus} objects can be used to represent the status
of a given document. A core requirement is that the different status
records should be persistent across different compilation runs. That
makes it necessary to store the intermediate states of an \code{R6ClassDocuStatus}
object in a file. Given that requirement we must have methods for
reading status information from a file and for writing status information
to a file. Furthermore, we need a method to add a document status record
and we must be able to display all document status records as a table.
}
\section{Fields}{

\describe{
\item{\code{version}}{current version of the document}

\item{\code{status_colnames}}{vector of column names shown in the document table}

\item{\code{status_history}}{dataframe with the document history read from the history file}

\item{\code{history_file}}{name of the file containing the document history}
}}
\section{Public Methods}{

\describe{
  \item{\code{new()}}{This method instantiates an object of class R6ClassDocuStatus}
  \item{\code{initialize(psFormat)}}{Initialization after creating the instance. Reads history file if it exists}
  \item{\code{include_doc_stat(psTitle)}}{Saves updated document status to the status file.
              Write section header psTitle for document status and write
              markdown table containing the document status.}
}
}

\section{Private Methods}{

\describe{
  \item{\code{writeStatusToFile()}}{Writes current status and history
              to a tab-separated file. Tab-separated format is chosen, because TAB-characters
              are less likely to occur in any of the table fields. File encoding is set
              to "UTF-8" in order to preserve German Umlauts}
  \item{\code{readStatusFromFile()}}{Document status history is read
              from the history file.}
  \item{\code{readCsv2StatusFromFile()}}{Reading method for old csv2
              formatted status history files. This is mainly used for converting history
              files from old csv2 format to new tab-separated format.}
  \item{\code{knitr_kable}}{Complete status history is written to a table using
        \code{knitr::kable} to convert the dataframe into a markdown-table}
  \item{\code{auto_increment()}}{Autoincrementing minor version number}
}
}
\examples{
r6objDocuStat <- R6ClassDocuStatus$new()
r6obj_docstat$set_current_status(psVersion = "0.0.908",
                                 psStatus  = "New version of setting document status",
                                 psProject = "rmddochelper")
\dontrun{
r6objDocuStat$writeStatusToFile()
r6objDocuStat$knitr_kable()
}
}
\keyword{datasets}

