% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmddochelper_table_abbrev_r6class.R
\docType{class}
\name{R6ClassTableAbbrev}
\alias{R6ClassTableAbbrev}
\title{R6 Class To Model Abbreviation Table Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassTableAbbrev$new()
}
\value{
R6 Object of type \code{\link{R6ClassTableAbbrev}} with
fields and methods to collect abbreviations and generate a table
of abbreviations
}
\description{
Abbreviation tables are tables with two columns one of which
contains the abbreviations and the second column contains the
meaning of the abbreviations. The tables are specific for a
given document. The R6 class defined here provides functionlity
to automatically collect all abbreviations together with the
associated meanings and to generate the resulting table of
abbreviations. Method \code{add_abbrev} is used to add new
abbreviations. At the end of a document the list of collected
abbreviations together with their meanings is written to a file.
This file is the source for automatically generating the table
of abbreviations. Since, we want to allow for including the
table of abbreviations at any position within the document,
two compilation runs are required when new abbreviations are
added to the list of abbreviations. The first run collects
all abbreviations and writes them to the file and the second
run is used to generate the table of abbreviations.
}
\section{Fields}{

\describe{
\item{\code{dfAbbrTable}}{dataframe with abbreviations}

\item{\code{colHeader}}{column header for table of abbreviations}

\item{\code{sAbbrFile}}{name of the file where abbreviations are written to}

\item{\code{sAbbrTitle}}{title to be used in document above table of abbreviations}
}}
\section{Public methods}{

\describe{
  \item{\code{new()}}{Instantiation of R6 object of class R6ClassTableAbbrev}
  \item{\code{setColHeader}}{setter for column header}
  \item{\code{setAbbrFile}}{setter for abbreviation file}
  \item{\code{setAbbrTitle}}{setter for abbreviation title}
  \item{\code{add_abbrevpsAbbrev, psMeaning, pbquote = TRUE, pbOut = TRUE}}
       {add new pair of abbreviation(psAbbrev) and meaning(psMeaning) to list of abbreviations.
        Flag pbQuote determines whether abbreviation and meaning are enclosed in quotes. Flag
        pbOut determines whether abbreviation and meaning should be written to Rmarkdown source file.}
  \item{\code{writeToTsvFile}}{Write list of abbreviations in tab-separated format to file}
  \item{\code{include_abbr_table}}{Include table of abbreviations in a document}
  \item{\code{is_empty_abbr}}{Check whether list of abbreviations is empty}
}
}
\examples{
r6objAbr <- R6ClassTableAbbrev$new()
r6objAbr$setAbbrFile(psAbbrFile = "ABBREVIATIONS")
r6objAbr$add_abbrev(psAbbrev = "Abr", psMeaning = "Abbreviation")
\dontrun{
r6objAbr$writeToTsvFile()
r6objAbr$include_abbr_table()
}
}
\keyword{datasets}

